requirements.txt
Flask>=3.0,<4
Werkzeug>=3.0,<4
python-dotenv>=1.0,<2
openai>=1.30,<2
reportlab>=3.6,<5
pdfplumber>=0.11,<1
